Add an Offline ReverseGeocoder
This operates by downloading a 6.7mb zip via cmake, unzips it (21mb) and
installs it. We then use a kd-tree to load all the data and perform
lookups. On my system it takes about 800 msecs to generate the kd-tree,
and near instant lookups (<1msec).

This OfflineReverseGeoCoder is not being used in Koko with this commit.